DTS-67275 | D360 24.0: Upgrade to WildFly 27 (Nov-2022)
Second pass at converting TEIID from the JAVAX to the JAKARTA EE.
Most of the build is working except for the following:
Olingo was removed as it has not yet been migrated to Jakarta EE.
Infinispan is giving issues.
Swagger needs to be migrated from the io.swagger modules to the io.swagger.v3:*-jakarta modules.