DTS-75885 | Seeming idle SQ server has 100% CPU usage TEIID-3106 | BufferManager Cleaner consuming 90% of total CPU time (revisited)
After analyzing a HEAP dump when the server was in this state it was found that the eviction queue had a single CacheEntry instance present. This object was not in the memoryEntries map and ageOut was true due to the call coming from the Cleaner.run() second eviction call. This left the BufferManagerImpl to spin in this while loop unable to break out. This change will track the previously audited CacheEntry instance and when it is the same instance as the CacheEntry being analyzed then we will break out of this loop.